Require auth before contacting group leader
