+++ components/GroupsResultsList/index.js
@@ -1,0 +1,3 @@
+import GroupsResultsList from './GroupsResultsList';
+
+export default GroupsResultsList;
